database changed from redis to couchbase

+++ basket/application/src/main/java/com/trendyol/basket/application/converter/impl/BasketInfoDtoConverter.java
@@ -5,6 +5,7 @@ import com.trendyol.basket.application.model.dto.BasketInfoDTO;
 import com.trendyol.basket.domain.entity.BasketInfo;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
 import java.util.stream.Collectors;
 
 @Component
@@ -21,8 +22,8 @@ public class BasketInfoDtoConverter implements Converter<BasketInfo, BasketInfoD
         var basketInfoDTO = new BasketInfoDTO();
         var campaignDTOs = campaignDtoConverter.convert(basketInfo.getCampaigns());
         basketInfoDTO.setCampaignDTOs(campaignDTOs);
-        basketInfoDTO.setSubTotal(basketInfo.getSubTotal());
-        basketInfoDTO.setGrandTotal(basketInfo.getGrandTotal());
+        basketInfoDTO.setSubTotal(BigDecimal.valueOf(basketInfo.getSubTotal()));
+        basketInfoDTO.setGrandTotal(BigDecimal.valueOf(basketInfo.getGrandTotal()));
         return basketInfoDTO;
     }
 }
