+++ basket/application/src/main/java/com/trendyol/basket/application/converter/impl/BasketItemDtoConverter.java
@@ -6,6 +6,8 @@ import com.trendyol.basket.application.model.dto.BasketItemDTO;
 import com.trendyol.basket.domain.entity.BasketItem;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
+
 @Component
 public class BasketItemDtoConverter implements Converter<BasketItem, BasketItemDTO> {
 
@@ -13,8 +15,8 @@ public class BasketItemDtoConverter implements Converter<BasketItem, BasketItemD
     public BasketItemDTO convert(BasketItem productInfo) {
         var productInfoDTO = new BasketItemDTO();
         productInfoDTO.setId(productInfo.getProductId());
-        productInfoDTO.setPrice(productInfo.getPrice());
-        productInfoDTO.setOldPrice(productInfo.getOldPrice());
+        productInfoDTO.setPrice(BigDecimal.valueOf(productInfo.getPrice()));
+        productInfoDTO.setOldPrice(BigDecimal.valueOf(productInfo.getOldPrice()));
         productInfoDTO.setQuantity(productInfo.getQuantity());
         productInfoDTO.setImageUrl(productInfo.getImageUrl());
         productInfoDTO.setTitle(productInfo.getTitle());
